+++ examples/text/text_draw_3d.c
@@ -152,13 +152,13 @@ int main(void)
             if (IsFileExtension(droppedFiles.paths[0], ".ttf"))
             {
                 UnloadFont(font);
-                font = LoadFontEx(droppedFiles.paths[0], fontSize, 0, 0);
+                font = LoadFontEx(droppedFiles.paths[0], (int)fontSize, 0, 0);
             }
             else if (IsFileExtension(droppedFiles.paths[0], ".fnt"))
             {
                 UnloadFont(font);
                 font = LoadFont(droppedFiles.paths[0]);
-                fontSize = font.baseSize;
+                fontSize = (float)font.baseSize;
             }
             
             UnloadDroppedFiles(droppedFiles);    // Unload filepaths from memory
@@ -742,7 +742,7 @@ static Vector3 MeasureTextWave3D(Font font, const char* text, float fontSize, fl
 static Color GenerateRandomColor(float s, float v)
 {
     const float Phi = 0.618033988749895f; // Golden ratio conjugate
-    float h = GetRandomValue(0, 360);
+    float h = (float)GetRandomValue(0, 360);
     h = fmodf((h + h*Phi), 360.0f);
     return ColorFromHSV(h, s, v);
 }
